+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [x] [#536 - Statusbar background color change delayed on application start](https://github.com/scriptPilot/app-framework/issues/536)
 - [x] [#543 - Statusbar shown on initial native app start, but configured as false](https://github.com/scriptPilot/app-framework/issues/543)
 - [x] [#554 - Vue warning due to material icon workaround for Android 4](https://github.com/scriptPilot/app-framework/issues/554)
+- [x] [#558 - $root.frameworkVersion / $root.version show previous version](https://github.com/scriptPilot/app-framework/issues/558)
 
 ### Improvements